moved ordertype from store to reservation
