+++ src/main/java/zupzup/back_end/reservation/dto/OrderDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.modelmapper.ModelMapper;
 import zupzup.back_end.reservation.domain.Order;
-import zupzup.back_end.store.domain.type.OrderStatus;
+import zupzup.back_end.reservation.domain.type.OrderStatus;
 import zupzup.back_end.reservation.dto.request.OrderRequestDto;
 
 import java.util.List;
